moved DelegateOnHost
